Agent Innovation Challenge/比赛提交（Competition Submission）/01灵犀-灵犀相册 Agent/Competition Code/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Vector Database
qdrant-client>=1.7.0

# Deep Learning
torch>=2.0.0
transformers>=4.40.0
pillow>=10.0.0
numpy>=1.24.0

# Qwen-VL Utils
qwen-vl-utils>=0.0.8

# HTTP Client (for future extensions)
httpx>=0.25.0

# File handling
python-multipart>=0.0.6
aiofiles>=23.2.0

openjiuwen>=0.1.3